Fix CSS counter() function resolution in ::before/::after pseudo-elements
Refactored counter-reset/counter-increment/counter-set handling to fix cumulative counter values:
- Extract parse_counter_decl_list() helper to parse counter declarations consistently
- counter-increment is now purely cumulative and never restored/popped
- counter-reset and counter-set push/pop scope stack for nesting support
- Introduced CounterScopeGuard RAII struct to ensure pop_css_counter_scopes() is always called

Key insight: counter-increment should be cumulative across siblings (within a single reset scope), while counter-reset/counter-set establish new scopes that can be nested. The fix ensures:
1. counter-reset: resets to 0 (or specified value), pushes previous value to scope stack
2. counter-increment: adds to current value, no scope manipulation
3. On scope exit: counter-reset/counter-set pop and restore previous values, but counter-increment increments persist

This allows correct behavior for:
  .list { counter-reset: item; }
  .list li { counter-increment: item; }
  .list li::before { content: counter(item) ". "; }
Each li will now display 1, 2, 3, ... correctly.

+++ vibrowser/src/shell/render_view.mm
@@ -708,7 +708,7 @@ struct TextRegion {
         if (std::abs(scrollAmountY) > kScrollDeltaEpsilon) {
             float maxScrollY = std::max(0.0f, scrollTarget->scroll_content_height - scrollTarget->geometry.height);
             float prevScroll = scrollTarget->scroll_top;
-            scrollTarget->scroll_top = std::max(0.0f, std::min(scrollTarget->scroll_top + scrollAmountY, maxScrollY));
+            scrollTarget->scroll_top = std::max(0.0f, std::min(static_cast<float>(scrollTarget->scroll_top + scrollAmountY), static_cast<float>(maxScrollY)));
             appliedDeltaY = scrollTarget->scroll_top - prevScroll;
         }
 
@@ -716,7 +716,7 @@ struct TextRegion {
         if (std::abs(scrollAmountX) > kScrollDeltaEpsilon) {
             float maxScrollX = std::max(0.0f, scrollTarget->scroll_content_width - scrollTarget->geometry.width);
             float prevScroll = scrollTarget->scroll_left;
-            scrollTarget->scroll_left = std::max(0.0f, std::min(scrollTarget->scroll_left + scrollAmountX, maxScrollX));
+            scrollTarget->scroll_left = std::max(0.0f, std::min(static_cast<float>(scrollTarget->scroll_left + scrollAmountX), static_cast<float>(maxScrollX)));
             appliedDeltaX = scrollTarget->scroll_left - prevScroll;
         }
